refactor: split DownloadPage responsibilities
